pyproject.toml
[project]
name = "snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "onnxruntime>=1.24.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pygame>=2.6.1",
    "rembg>=2.0.72",
]
